pyproject.toml
[tool.poetry]
name = "final-project-sample"
version = "0.1.0"
description = ""
authors = ["Mjumbe Poe <mjumbewu@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
sqlalchemy-bigquery = "^1.2.1"
SQLAlchemy = "^1.4.26"
google-cloud-storage = "^1.42.3"
google-cloud-bigquery-storage = "^2.10.0"
geopandas = "^0.10.2"
pandas-gbq = "^0.16.0"
Jinja2 = "^3.0.3"

[tool.poetry.dev-dependencies]
jupyterlab = "^3.2.2"
matplotlib = "^3.4.3"
branca = "^0.4.2"
folium = "^0.12.1"
mapclassify = "^2.4.3"
python-dotenv = "^0.19.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
